feat(story): delete story, make mandatory fields non nullable
Initially the idea was to force audience, type and copyright to non nullable

But I had a few stories to delete on the server, so I wrote down the feature.
Basically removes the stories, and its associations to genres, TW and collaborators.
There is some thinking to do about consequences of story deletion, but I keep it for later